Add cpp_rule_preprocessor main
