+++ cpp2rust/rules_dir.h
@@ -1,0 +1,12 @@
+#pragma once
+
+// Copyright (c) 2022-present INESC-ID.
+// Distributed under the MIT license that can be found in the LICENSE file.
+
+#include <string>
+
+namespace cpp2rust {
+
+bool ResolveRulesDir(std::string &out);
+
+} // namespace cpp2rust
